把报警能去的都去掉了；增加了HMI部分的VRAM分配；编写了将时间写入模型的代码

+++ device/dev_lcd.h
@@ -71,7 +71,7 @@ typedef struct{
 	void ( *getScrnSize)( uint16_t *xsize, uint16_t *ysize);
 	void ( *picture)( int x1, int y1, char num);
 	void ( *cutPicture)( short x1, short y1, short num, short px1, short py1, short w, short h);
-	void ( *bkPicture)( char m, short x1, short y1, char num);
+	void ( *bkPicture)( char m, int x1, int y1, char num);
 	void	(*done)(void);
 	void	(*icon)(int x1, int y1, char num, int xn, int yn, int n);
 }I_dev_lcd;
